:memo: 중복 제거
